Add Windows ocaml-system- + base-system- packages
Add the ocaml-system-mingw and ocaml-system-msvc packages to specify
either the mingw-w64 or MSVC ports when compiling OCaml and the
base-system-mingw and base-system-msvc packages to be used for the
dependency graph.

The intention is to complete this for non-Windows systems, but as there
will always be a chance of an unknown system, base-system-other is added
to be used for all the other ports. This package is significant, as
compiler packages must always install a base-system- package (or the
user could attempt to install another incorrect one).