Add Windows arch- + all host-arch- packages
A package is available for each supported OCaml architecture.
host-arch-unknown is added to ensure that each compiler package is
always able to install one of these packages.

The intention is to complete this for non-Windows systems, and in
particular to ensure that this is fully compatible with
ocaml-option-32bit. For now, only arch-x86_32 and arch-x86_64 are
available, as these are the two supported Windows architectures.